README: move copyright related information into a single spot
Some build system may want to cite or extract license information from a
project's sources. Referencing two separate spots within the same file
may not be possible.

Move the "Copyright and license" section to the top of the file, such
that an adjacent block of lines spans all related information. This is
very desirable for bitbake recipes in a poky layer.